backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
requests==2.31.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
urllib3==2.1.0
python-dotenv==1.0.0
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3